requirements.txt
# Xoul Agent — VM 내부 의존성
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.0.0
holidays>=0.40
beautifulsoup4>=4.12.0
openai>=1.0.0
